# openapi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "electrostore",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "{baseUrl}",
      "description": "URL de base personnalisable",
      "variables": {
        "baseUrl": {
          "default": "https://exemple.com",
          "description": "URL de base personnalisée",
          "examples": [
            "https://api.exemple.com",
            "https://staging-api.exemple.com",
            "http://localhost:8080"
          ]
        }
      }
    }
  ],
  "host": "{baseUrl}",
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/api/item/{id_item}/box/{id_box}": {
      "get": {
        "tags": [
          "ItemBox"
        ],
        "parameters": [
          {
            "name": "id_item",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_box",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadItemBoxDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadItemBoxDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadItemBoxDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ItemBox"
        ],
        "parameters": [
          {
            "name": "id_item",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_box",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemBoxByItemDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemBoxByItemDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemBoxByItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadItemBoxDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadItemBoxDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadItemBoxDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ItemBox"
        ],
        "parameters": [
          {
            "name": "id_item",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_box",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemBoxDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemBoxDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemBoxDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadItemBoxDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadItemBoxDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadItemBoxDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ItemBox"
        ],
        "parameters": [
          {
            "name": "id_item",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_box",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/item/{id_item}/tag/{id_tag}": {
      "get": {
        "tags": [
          "ItemTag"
        ],
        "parameters": [
          {
            "name": "id_item",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_tag",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadItemTagDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadItemTagDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadItemTagDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ItemTag"
        ],
        "parameters": [
          {
            "name": "id_item",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_tag",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadItemTagDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadItemTagDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadItemTagDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ItemTag"
        ],
        "parameters": [
          {
            "name": "id_item",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_tag",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/store/{id_store}/box/{id_box}": {
      "get": {
        "tags": [
          "StoreBox"
        ],
        "parameters": [
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_box",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "StoreBox"
        ],
        "parameters": [
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_box",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBoxByStoreDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBoxByStoreDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBoxByStoreDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "StoreBox"
        ],
        "parameters": [
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_box",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/store/{id_store}/box/{id_box}/tag/{id_tag}": {
      "get": {
        "tags": [
          "StoreBoxTag"
        ],
        "parameters": [
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_box",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_tag",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxTagDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxTagDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxTagDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "StoreBoxTag"
        ],
        "parameters": [
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_box",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_tag",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxTagDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxTagDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxTagDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "StoreBoxTag"
        ],
        "parameters": [
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_box",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_tag",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/store/{id_store}/tag/{id_tag}": {
      "get": {
        "tags": [
          "StoreTag"
        ],
        "parameters": [
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_tag",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreTagDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreTagDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreTagDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "StoreTag"
        ],
        "parameters": [
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_tag",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreTagDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreTagDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreTagDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "StoreTag"
        ],
        "parameters": [
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_tag",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/tag/{id_tag}": {
      "get": {
        "tags": [
          "Tag"
        ],
        "parameters": [
          {
            "name": "id_tag",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadTagDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadTagDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadTagDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Tag"
        ],
        "parameters": [
          {
            "name": "id_tag",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTagDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTagDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadTagDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadTagDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadTagDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tag"
        ],
        "parameters": [
          {
            "name": "id_tag",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/tag/{id_tag}/box": {
      "get": {
        "tags": [
          "TagBox"
        ],
        "parameters": [
          {
            "name": "id_tag",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Total-Count": {
                "description": "Total number of items",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadBoxTagDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadBoxTagDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadBoxTagDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TagBox"
        ],
        "parameters": [
          {
            "name": "id_tag",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxTagDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxTagDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxTagDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/tag/{id_tag}/box/{id_box}": {
      "get": {
        "tags": [
          "TagBox"
        ],
        "parameters": [
          {
            "name": "id_tag",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_box",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxTagDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxTagDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxTagDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TagBox"
        ],
        "parameters": [
          {
            "name": "id_tag",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_box",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxTagDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxTagDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxTagDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TagBox"
        ],
        "parameters": [
          {
            "name": "id_tag",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_box",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateItemBoxByItemDto": {
        "required": [
          "qte_item_box",
          "seuil_max_item_item_box"
        ],
        "type": "object",
        "properties": {
          "qte_item_box": {
            "type": "integer",
            "format": "int32"
          },
          "seuil_max_item_item_box": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ReadBoxDto": {
        "type": "object",
        "properties": {
          "id_box": {
            "type": "integer",
            "format": "int32"
          },
          "xstart_box": {
            "type": "integer",
            "format": "int32"
          },
          "ystart_box": {
            "type": "integer",
            "format": "int32"
          },
          "xend_box": {
            "type": "integer",
            "format": "int32"
          },
          "yend_box": {
            "type": "integer",
            "format": "int32"
          },
          "id_store": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ReadBoxTagDto": {
        "type": "object",
        "properties": {
          "id_tag": {
            "type": "integer",
            "format": "int32"
          },
          "id_box": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ReadItemBoxDto": {
        "type": "object",
        "properties": {
          "id_box": {
            "type": "integer",
            "format": "int32"
          },
          "id_item": {
            "type": "integer",
            "format": "int32"
          },
          "qte_item_box": {
            "type": "integer",
            "format": "int32"
          },
          "seuil_max_item_item_box": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ReadItemTagDto": {
        "type": "object",
        "properties": {
          "id_item": {
            "type": "integer",
            "format": "int32"
          },
          "id_tag": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateItemBoxDto": {
        "type": "object",
        "properties": {
          "new_id_box": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "qte_item_box": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "seuil_max_item_item_box": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [
        
      ]
    }
  ]
}